daemon: gate done state on uncordon completion
The "complete update" workflow in the daemon today is quite convoluted,
which can cause the daemon to never uncordon the node, if the first
attempt reaches the timeout.

This problem is mitigated in 4.11+ since the drain is conducted by the
controller, but in 4.10 and prior the MCD is still doing the
drain/uncordon. We should probably rework this logic entirely, but for
4.10, let's take the less risky/less effort route of just fixing the one
issue above.